+++ embedded-graphics/src/drawable.rs
@@ -24,11 +24,14 @@ use crate::{geometry::Point, pixelcolor::PixelColor, DrawTarget};
 ///     text: &'a str,
 /// }
 ///
-/// impl<'a, C: 'a> Drawable<C> for &Button<'a, C>
+/// impl<C> Drawable<C> for Button<'_, C>
 /// where
 ///     C: PixelColor + From<BinaryColor>,
 /// {
-///     fn draw<D: DrawTarget<Color = C>>(self, display: &mut D) -> Result<(), D::Error> {
+///     fn draw<D>(&self, display: &mut D) -> Result<(), D::Error>
+///     where
+///         D: DrawTarget<Color = C>,
+///     {
 ///         Rectangle::new(self.top_left, self.size)
 ///             .into_styled(PrimitiveStyle::with_fill(self.bg_color))
 ///             .draw(display)?;
@@ -61,7 +64,9 @@ where
     C: PixelColor,
 {
     /// Draw the graphics object using the supplied DrawTarget.
-    fn draw<D: DrawTarget<Color = C>>(self, display: &mut D) -> Result<(), D::Error>;
+    fn draw<D>(&self, display: &mut D) -> Result<(), D::Error>
+    where
+        D: DrawTarget<Color = C>;
 }
 
 /// A single pixel.
@@ -105,18 +110,11 @@ impl<C> Drawable<C> for Pixel<C>
 where
     C: PixelColor,
 {
-    fn draw<D: DrawTarget<Color = C>>(self, display: &mut D) -> Result<(), D::Error> {
-        display.draw_iter(core::iter::once(self))
-    }
-}
-
-impl<C, T> Drawable<C> for &mut T
-where
-    C: PixelColor,
-    T: Iterator<Item = Pixel<C>>,
-{
-    fn draw<D: DrawTarget<Color = C>>(self, display: &mut D) -> Result<(), D::Error> {
-        display.draw_iter(self)
+    fn draw<D>(&self, display: &mut D) -> Result<(), D::Error>
+    where
+        D: DrawTarget<Color = C>,
+    {
+        display.draw_iter(core::iter::once(*self))
     }
 }