+++ embedded-graphics/src/pixel_iterator.rs
@@ -1,0 +1,28 @@
+//! Pixel iterator
+
+use crate::drawable::Pixel;
+use crate::{pixelcolor::PixelColor, DrawTarget};
+
+/// Extension trait for pixel iterators.
+pub trait PixelIteratorExt<C>
+where
+    C: PixelColor,
+{
+    /// Draws the pixel iterator to a draw target.
+    fn draw<D>(self, target: &mut D) -> Result<(), D::Error>
+    where
+        D: DrawTarget<Color = C>;
+}
+
+impl<I, C> PixelIteratorExt<C> for I
+where
+    C: PixelColor,
+    I: Iterator<Item = Pixel<C>>,
+{
+    fn draw<D>(self, target: &mut D) -> Result<(), D::Error>
+    where
+        D: DrawTarget<Color = C>,
+    {
+        target.draw_iter(self)
+    }
+}