+++ embedded-graphics/src/prelude.rs
@@ -5,6 +5,7 @@ pub use crate::{
     fonts::Font,
     geometry::{Angle, AngleUnit, Dimensions, Point, Size},
     image::{ImageDimensions, IntoPixelIter},
+    pixel_iterator::PixelIteratorExt,
     pixelcolor::{raw::RawData, GrayColor, IntoStorage, PixelColor, RgbColor},
     primitives::{ContainsPoint, Primitive},
     transform::Transform,