+++ embedded-graphics/src/primitives/ellipse/styled.rs
@@ -90,11 +90,14 @@ where
     }
 }
 
-impl<'a, C: 'a> Drawable<C> for &Styled<Ellipse, PrimitiveStyle<C>>
+impl<C> Drawable<C> for Styled<Ellipse, PrimitiveStyle<C>>
 where
     C: PixelColor,
 {
-    fn draw<D: DrawTarget<Color = C>>(self, display: &mut D) -> Result<(), D::Error> {
+    fn draw<D>(&self, display: &mut D) -> Result<(), D::Error>
+    where
+        D: DrawTarget<Color = C>,
+    {
         display.draw_iter(self)
     }
 }