+++ embedded-graphics/src/primitives/triangle/styled.rs
@@ -74,11 +74,14 @@ where
     }
 }
 
-impl<'a, C: 'a> Drawable<C> for &Styled<Triangle, PrimitiveStyle<C>>
+impl<C> Drawable<C> for Styled<Triangle, PrimitiveStyle<C>>
 where
     C: PixelColor,
 {
-    fn draw<D: DrawTarget<Color = C>>(self, display: &mut D) -> Result<(), D::Error> {
+    fn draw<D>(&self, display: &mut D) -> Result<(), D::Error>
+    where
+        D: DrawTarget<Color = C>,
+    {
         display.draw_iter(self)
     }
 }